refactor: move non-public API into a sub-namespace called `_internal`, so that its reliability is obvious.
