+++ src/_internal/README.md
@@ -1,0 +1,5 @@
+# INTERNAL
+
+All in here is to be treated as non-public API.  
+All code in here might change without any notice during version upgrades.
+No backwards compatibility is guaranteed.
