+++ tests/Functional/PluginTest.php
@@ -26,8 +26,8 @@ namespace CycloneDX\Tests\Functional;
 use Composer\Plugin\Capability\CommandProvider;
 use Composer\Plugin\Capable;
 use Composer\Plugin\PluginInterface;
-use CycloneDX\Composer\MakeBom\Command;
-use CycloneDX\Composer\MakeBom\Options;
+use CycloneDX\Composer\_internal\MakeBom\Command;
+use CycloneDX\Composer\_internal\MakeBom\Options;
 use CycloneDX\Composer\Plugin;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\CoversNothing;
